Split JPEG.kt and add tests for encoder and decoder
